+++ packages/loopring_v3/contracts/impl/libexchange/ExchangeDeposits.sol
@@ -77,7 +77,7 @@ library ExchangeDeposits
         view
         returns (
           bytes32 accumulatedHash,
-          uint256 accumulatedFee,
+          uint    accumulatedFee,
           uint32  timestamp
         )
     {
